refactor importer into core libs & add json decode error check
